fix(DF-92): 검색 결과 클릭시 해당 위치로 이동되지 않는 문제 해결 - 해결방법: onOutsideClick 속성을 검색창이 아닌 검색결과창에 적용함
